Extract server logic to new class
