+++ lib/slsk-client.js
@@ -11,134 +11,48 @@ const debugPeer = require('debug')('slsk:peer:i')
 const Messages = require('./messages.js')
 const MessageFactory = require('./message-factory.js')
 const downloadPeerFile = require('./download-peer-file.js')
+const Server = require('./server.js')
 let stack = require('./stack')
 
-// TODO make this to store, to be required by other modules
-let client
+let server
 let peers = {}
 
 class SlskClient extends EventEmitter {
   init (cb) {
     debug('Init client')
+    server = new Server(cb)
 
-    client = net.createConnection({
-      host: 'server.slsknet.org',
-      port: 2242
-    }, cb)
-
-    let msgs = new Messages()
-
-    client.on('error', err => {
-      cb(new Error(err.message))
-    })
-
-    msgs.on('message', msg => {
-      let size = msg.int32()
-      if (size < 4) return
-      let code = msg.int32()
-      switch (code) {
-        case 1: {
-          debug(`Login Response`)
-          if (!stack.login) return
-          let success = msg.int8()
-          if (success === 1) {
-            stack.login()
-            delete stack.login
-          } else {
-            let reason = msg.str()
-            stack.login(new Error(reason))
-            delete stack.login
-          }
-          break
-        }
-        case 18: {
-          debug(`recv ConnectToPeer`)
-          this.connectToPeer(msg)
-          break
-        }
-        case 64: {
-          debug(`recv RoomList ${msg.data.length}`)
-          let nbRooms = msg.int32()
-          let rooms = []
-          for (let i = 0; i < nbRooms; i++) {
-            rooms.push({
-              name: msg.str()
-            })
-          }
-          for (let i = 0; i < nbRooms; i++) {
-            rooms[i].users = msg.int32()
-          }
-          break
-        }
-        case 69: {
-          let number = msg.int32()
-          debug(`there are ${number} PrivilegedUsers. msg length: ${msg.data.length}`)
-          break
-        }
-        case 83: {
-          let number = msg.int32()
-          debug(`ParentMinSpeed is ${number}. msg length: ${msg.data.length}`)
-          break
-        }
-        case 84: {
-          let number = msg.int32()
-          debug(`ParentSpeedRatio is ${number}. msg length: ${msg.data.length}`)
-          break
-        }
-        case 104: {
-          let number = msg.int32()
-          debug(`Whishlist interval is ${number}. msg length: ${msg.data.length}`)
-          break
-        }
-        default: {
-          debug(`unknown srv message code: ${code} length: ${msg.data.length}`)
-          if (code > 1002) {
-            msgs.reset()
-          }
-        }
-      }
-    })
-
-    client.on('data', data => {
-      msgs.write(data)
+    server.on('connectToPeer', peer => {
+      this.connectToPeer(peer)
     })
   }
 
-  connectToPeer (msg) {
-    let user = msg.str()
-    let type = msg.str()
-    let ip = []
-    for (let i = 0; i < 4; i++) {
-      ip.push(msg.int8())
-    }
-    let host = ip[3] + '.' + ip[2] + '.' + ip[1] + '.' + ip[0]
-    let port = msg.int32()
-    let token = msg.readRawHexStr(4)
-    debugPeer(`connectToPeer ${user} ${host} ${port} ${token} ${type}`)
+  connectToPeer (peer) {
+    debugPeer(`connectToPeer ${peer.user} ${peer.host} ${peer.port} ${peer.token} ${peer.type}`)
 
-    if (type === 'F') {
-      downloadPeerFile(host, port, token, user, false)
+    if (peer.type === 'F') {
+      downloadPeerFile(peer.host, peer.port, peer.token, peer.user, false)
       return
     }
 
     let conn = net.createConnection({
-      host,
-      port
+      host: peer.host,
+      port: peer.port
     }, () => {
-      let buf = Buffer.from('05000000' + '00' + token, 'hex')
+      let buf = Buffer.from('05000000' + '00' + peer.token, 'hex')
       conn.write(buf)
-      peers[user] = conn
+      peers[peer.user] = conn
     })
 
     conn.on('error', error => {
-      debug(`${user} error ${error.code}`)
+      debug(`${peer.user} error ${error.code}`)
     })
 
     let msgs = new Messages()
 
     msgs.on('message', msg => {
       let size = msg.int32()
-      debugPeer(`${user} size: ${size}`)
+      debugPeer(`${peer.user} size: ${size}`)
       if (size <= 4) return
 
       let code = msg.int32()
@@ -185,7 +99,7 @@ class SlskClient extends EventEmitter {
           let file = msg.str()
           debugPeer(`recv TransferRequest ${file}`)
           stack.downloadTokens[token] = {
-            user,
+            user: peer.user,
             file
           }
           if (dir === 1) {
@@ -211,18 +125,18 @@ class SlskClient extends EventEmitter {
             delete stack.downloadTokens[token] // avoid memory leak
           } else if (allowed === 1) {
             debugPeer(`Directly allowed. Connecting to peer with PeerInit + ${token}`)
-            downloadPeerFile(host, port, token, user, true)
+            downloadPeerFile(peer.host, peer.port, peer.token, peer.user, true)
           }
           break
         }
         case 46: {
           let filename = msg.str()
           debugPeer(`UploadFailed ${filename}`)
-          let down = stack.download[user + '_' + filename]
+          let down = stack.download[peer.user + '_' + filename]
           if (down && typeof down.cb === 'function') {
             down.cb(new Error('Peer error'))
           } else {
-            debugPeer(`Cannot cb for ${user} ${filename}`)
+            debugPeer(`Cannot cb for ${peer.user} ${filename}`)
           }
           break
         }
@@ -239,10 +153,7 @@ class SlskClient extends EventEmitter {
 
   login (credentials, cb) {
     stack.currentLogin = credentials.user
-    let msg = MessageFactory
-      .to.server
-      .login(credentials)
-    client.write(msg.getBuff())
+    server.login(credentials)
     stack.login = cb
   }
 
@@ -267,10 +178,7 @@ class SlskClient extends EventEmitter {
       query: obj.req
     }
 
-    let msg = MessageFactory
-      .to.server
-      .fileSearch(obj.req, token)
-    client.write(msg.getBuff())
+    server.fileSearch(obj.req, token)
   }
 
   download (obj, cb, stream) {
@@ -315,7 +223,7 @@ class SlskClient extends EventEmitter {
   }
 
   destroy () {
-    client.destroy()
+    server.destroy()
     delete stack.login
   }
 }
